Resolving the datasets access againt the permissions server.
There are 2 cases: user is authenticated, or not.
We send an extra boolean to the database to distinguish the 2 cases.

In the case the user is authenticated, there is a token.
We send the token to the permissions server, along with the requested list.
* If the list is empty, the server returns _all_ the datasets the user has access to.
* If the list is not empty, the server keeps in the list only the ones accessible (and the list is therefore empty)

In the case there is no token, we don't contact the permissions server and send the requested datasets list to the databases.
Given that the boolean is false, the database will intersect the public datasets with the list.
If the intersection is empty, it's normal, the user is requesting something it doesn't have access to: no results.