Adding a response streamer.
At the moment, given a python object (say a dict),
it iterates through the keys and values (descending in the values if necessary),
and spits out the converted strings.

We pass those generated strings, to a StreamResponse.

This lowers the memory consumption as we won't need to create the python struct,
and we'll only need to create generator (in a later commit).